requirements.txt
# Core Framework
fastapi==0.125.0
uvicorn[standard]==0.32.1
streamlit==1.32.2

# AI & ML Core
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.63
langchain-groq==0.3.2
langchain-huggingface==0.2.0
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langchain-chroma==0.1.4
openai==1.57.4

# Vector & Search
chromadb==0.5.23
rank_bm25==0.2.2
sentence-transformers==3.3.1
faiss-cpu==1.9.0

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
markdown==3.7
beautifulsoup4==4.12.3
unstructured==0.16.9

# Database & Caching
sqlalchemy==2.0.36
redis==5.2.1
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0

# Monitoring & Analytics
prometheus-client==0.21.0
python-json-logger==3.2.1
sentry-sdk==2.19.2

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0
aiofiles==24.1.0
celery==5.4.0
httpx==0.28.1
tenacity==9.0.0

# Testing & Quality
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.8.4

# CORS & Middleware
python-cors==1.0.0
slowapi==0.1.9
